move infer back to CM module
