+++ compute_modules/bin/ontology/generate_metadata_config.py
