+++ compute_modules/bin/static_inference/infer.py
